converted otter generate to click
